requirements.txt
https://www.djangoproject.com/download/1.7c1/tarball/
djangorestframework==2.3.14
Markdown==2.4.1
django-filter==0.7
selenium==2.42.1
git+https://github.com/galuszkak/django-behave#egg=django-behave
behave==1.2.4
behaving==0.7
chaussette==1.1
django-allauth==0.17.0
oauthlib==0.6.3
parse==1.6.4
parse-type==0.3.4
python3-openid==3.0.4
requests==2.3.0
requests-oauthlib==0.4.1
six==1.7.3
splinter==0.6.0
django-crispy-forms==1.4.0